Move ModelMapper initialization to AppConfig
